Merge: Pushing sprint 1 bug-fixes to dev branch

See merge request polytechnique-montr-al/log2995/20261/equipe-310/LOG2995-310!32

+++ client/src/app/pages/admin-page/admin-page.component.ts
@@ -9,6 +9,7 @@ import { GameCard } from '@app/interfaces/gameCard';
 import { AdminGameService } from '@app/services/admin-game/admin-game.service';
 import { CommunicationService } from '@app/services/communication/communication.service';
 import { Subscription } from 'rxjs';
+import swal from 'sweetalert2';
 
 @Component({
   selector: 'app-admin-page',
@@ -29,13 +30,18 @@ export class AdminPageComponent implements OnInit, OnDestroy {
     private readonly adminGameService: AdminGameService,
     private readonly router: Router,
   ) {}
-
+  
   ngOnInit(): void {
     this.adminGameService.fetchAllGames().subscribe({
       next: (games) => this.adminGameService.setGames(games),
       error: (error: HttpErrorResponse) => {
         const errorMessage = error.error || 'Erreur lors de la récupération des jeux';
-        alert(`Erreur: ${errorMessage}`);
+        swal.fire({
+          title: `Erreur`,
+          text: `${errorMessage}`,
+          icon: 'error',
+          confirmButtonText: 'OK',
+        });
       },
     });
 
@@ -66,7 +72,7 @@ export class AdminPageComponent implements OnInit, OnDestroy {
 
   navigateToGameEditor(name: string): void {
     const game = this.games.find((g) => (g.name === name));
-    this.router.navigate(['/editor'], { state: { game, mode: 'edit' } });
+    this.router.navigate(['/editor', game?._id], { state: { game, mode: 'edit' } });
   }
 
   changeVisibility(name: string) {
@@ -75,20 +81,44 @@ export class AdminPageComponent implements OnInit, OnDestroy {
     this.communicationService.updateVisiblity(game).subscribe({
       error: (error: HttpErrorResponse) => {
         const errorMessage = error.error || 'Erreur lors du changement de visibilité';
-        alert(`Erreur: ${errorMessage}`);
+        swal.fire({
+          title: `Erreur`,
+          text: `${errorMessage}`,
+          icon: 'error',
+          confirmButtonText: 'OK',
+        });
       },
     });
   }
 
-  removeGame(name: string) {
+  async removeGame(name: string) {
+
     const game = this.games.find(g => g.name === name);
     if (!game) return;
 
-    this.communicationService.deleteGame(game._id).subscribe({
-      error: (error: HttpErrorResponse) => {
-        const errorMessage = error.error || 'Erreur lors de la suppression du jeu';
-        alert(`Erreur: ${errorMessage}`);
-      },
+    const result = await swal.fire({
+      title: 'Êtes-vous sûrs de vouloir supprimer ce jeu ?',
+      text: `Cette action est irréversible.`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Oui, supprimer',
+      cancelButtonText: 'Annuler',
     });
+
+    if (result.isConfirmed) {
+      this.communicationService.deleteGame(game._id).subscribe({
+        error: (error: HttpErrorResponse) => {
+          const errorMessage = error.error || 'Erreur lors de la suppression du jeu';
+          swal.fire({
+          title: `Erreur`,
+          text: `${errorMessage}`,
+          icon: 'error',
+          confirmButtonText: 'OK',
+        });
+        },
+      });
+    }
   }
 }
